move the contrib fetching code to it's own file
