+++ src/github_contrib_view/main.py
@@ -9,17 +9,12 @@ from bisect import bisect_left
 from datetime import datetime, timedelta, timezone
 from typing import NoReturn, Optional, TypedDict
 
-import requests
 import typer
 from rich import print as rprint
 
 from github_contrib_view.config import settings
-from github_contrib_view.constants import (
-    DAYS_PER_WEEK,
-    HTTP_OK,
-    QUERY,
-    REQUEST_TIMEOUT,
-)
+from github_contrib_view.constants import DAYS_PER_WEEK
+from github_contrib_view.contrib import get_github_contributions
 
 
 # Define a TypedDict for options
@@ -42,52 +37,6 @@ app = typer.Typer(
     add_completion=False,
     rich_markup_mode="rich",
 )
-
-
-def get_github_contributions(
-    username: str, token: str
-) -> dict[str, int] | None:
-    """Get GitHub contributions using GraphQL API.
-
-    You need a GitHub Personal Access Token
-    """
-    headers = {
-        "Authorization": f"Bearer {token}",
-        "Content-Type": "application/json",
-    }
-
-    response = requests.post(
-        "https://api.github.com/graphql",
-        json={"query": QUERY, "variables": {"username": username}},
-        headers=headers,
-        timeout=REQUEST_TIMEOUT,
-    )
-
-    data = response.json()
-
-    if data["data"]["user"] is None:
-        rprint(
-            f"[red]Error[/red] : User '{username}' cannot be found on GitHub, "
-            "[red]Exiting[/red]"
-        )
-        raise typer.Exit(code=1)
-
-    if response.status_code == HTTP_OK:
-        weeks = data["data"]["user"]["contributionsCollection"][
-            "contributionCalendar"
-        ]["weeks"]
-
-        contributions = {}
-        for week in weeks:
-            for day in week["contributionDays"]:
-                date = day["date"]
-                count = day["contributionCount"]
-                contributions[date] = count
-
-        return contributions
-
-    print(f"Error: {response.status_code}")
-    return None
 
 
 def get_symbol(count: int, *, use_ascii: bool = False) -> str:
